Fix(RHOAIENG-58501): address PR review comments for stop/retry and results UI
- Fix test assertions to match actual two-arg notification.success call
- Rename data-testid from run-canceled-label to run-status-label since
  it is used for both CANCELED and FAILED states
- Move terminate/retry handlers from pipeline_run_handler.go (singular)
  to pipeline_runs_handler.go (plural) in autorag for consistency with
  automl and existing handler conventions
- Fix pre-existing test assertion expecting 'Canceled' instead of 'CANCELED'
