+++ packages/autorag/frontend/src/app/components/run-results/AutoragResults.tsx
@@ -141,7 +141,7 @@ function AutoragResults(): React.JSX.Element {
                 <Label
                   variant="outline"
                   status={isCanceled ? 'warning' : 'danger'}
-                  data-testid="run-canceled-label"
+                  data-testid="run-status-label"
                 >
                   {pipelineRun.state}
                 </Label>
