+++ packages/autorag/frontend/src/app/pages/__tests__/AutoragResultsPage.spec.tsx
@@ -624,7 +624,10 @@ describe('AutoragResultsPage', () => {
       await userEvent.click(screen.getByTestId('confirm-stop-run-button'));
 
       await waitFor(() => {
-        expect(mockNotification.success).toHaveBeenCalledWith('Run stopped successfully');
+        expect(mockNotification.success).toHaveBeenCalledWith(
+          'Stop submitted successfully',
+          'The process is asynchronous and may take some time to take effect',
+        );
       });
     });
